+++ app/src/androidTest/java/com/rrajath/bloggo/ui/editor/EditorScreenTest.kt
@@ -59,7 +59,7 @@ class EditorScreenTest {
           onToast = {},
           onDeletePost = {},
           onMoveToInbox = {},
-          onPublish = { _, _ -> },
+          onPublish = { _, _, _ -> },
           onInsertImage = {},
           onPendingInsertConsumed = {},
         )